Refactor XPath Axes to a dedicated object available with FluentWebElement.axes()
